+++ src/app/loadout/loadout-edit/LoadoutEditBucket.m.scss.d.ts
@@ -10,6 +10,7 @@ interface CssExports {
   'itemCategory': string;
   'items': string;
   'itemsInCategory': string;
+  'missingItem': string;
   'showFashion': string;
   'unequipped': string;
 }