+++ src/app/loadout/loadout-edit/LoadoutEditBucket.tsx
@@ -1,8 +1,10 @@
 import { LoadoutParameters } from '@destinyitemmanager/dim-api-types';
 import ClosableContainer from 'app/dim-ui/ClosableContainer';
 import { t } from 'app/i18next-t';
+import ConnectedInventoryItem from 'app/inventory/ConnectedInventoryItem';
 import { InventoryBucket } from 'app/inventory/inventory-buckets';
 import { DimItem, PluggableInventoryItemDefinition } from 'app/inventory/item-types';
+import ItemPopupTrigger from 'app/inventory/ItemPopupTrigger';
 import { bucketsSelector } from 'app/inventory/selectors';
 import { LockableBucketHashes } from 'app/loadout-builder/types';
 import { DimLoadoutItem, Loadout } from 'app/loadout-drawer/loadout-types';
@@ -20,7 +22,6 @@ import { useSelector } from 'react-redux';
 import FashionDrawer from '../fashion/FashionDrawer';
 import { BucketPlaceholder } from '../loadout-ui/BucketPlaceholder';
 import { FashionMods } from '../loadout-ui/FashionMods';
-import ItemWithFashion from '../loadout-ui/ItemWithFashion';
 import LoadoutParametersDisplay from '../loadout-ui/LoadoutParametersDisplay';
 import { OptimizerButton } from '../loadout-ui/OptimizerButton';
 import styles from './LoadoutEditBucket.m.scss';
@@ -75,7 +76,6 @@ export default function LoadoutEditBucket({
             bucket={bucket}
             items={itemsByBucket[bucket.hash]}
             equippedItemIds={equippedItemIds}
-            modsForBucket={modsByBucket[bucket.hash] ?? emptyArray()}
             onClickPlaceholder={onClickPlaceholder}
             onClickWarnItem={onClickWarnItem}
             onRemoveItem={onRemoveItem}
@@ -147,17 +147,15 @@ function ItemBucket({
   items,
   equippedItemIds,
   equippedContent,
-  modsForBucket,
   onClickPlaceholder,
   onClickWarnItem,
   onRemoveItem,
   onToggleEquipped,
 }: {
   bucket: InventoryBucket;
-  items: DimItem[];
+  items: DimLoadoutItem[];
   equippedItemIds: Set<string>;
   equippedContent?: React.ReactNode;
-  modsForBucket: number[];
   onClickPlaceholder: (params: { bucket: InventoryBucket }) => void;
   onClickWarnItem: (item: DimItem) => void;
   onRemoveItem: (item: DimItem) => void;
@@ -186,19 +184,28 @@ function ItemBucket({
             className={clsx(styles.items, index === 0 ? styles.equipped : styles.unequipped)}
             key={index}
           >
-            {items.map((item, index) => (
+            {items.map((item) => (
               <ClosableContainer
                 key={item.id}
                 onClose={() => onRemoveItem(item)}
                 showCloseIconOnHover
               >
-                <ItemWithFashion
-                  item={item}
-                  applyFashion={showFashion && index === 0}
-                  modsForBucket={modsForBucket}
-                  onMissingItemClick={() => onClickWarnItem(item)}
-                  onDoubleClick={() => onToggleEquipped(item)}
-                />
+                <ItemPopupTrigger item={item} extraData={{ socketOverrides: item.socketOverrides }}>
+                  {(ref, onClick) => (
+                    <div
+                      className={clsx({
+                        [styles.missingItem]: item.owner === 'unknown',
+                      })}
+                    >
+                      <ConnectedInventoryItem
+                        item={item}
+                        innerRef={ref}
+                        onClick={item.owner === 'unknown' ? () => onClickWarnItem(item) : onClick}
+                        onDoubleClick={() => onToggleEquipped(item)}
+                      />
+                    </div>
+                  )}
+                </ItemPopupTrigger>
               </ClosableContainer>
             ))}
             {index === 0 && equippedContent}