+++ docs/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Fix applying D1 loadouts.
 * `inloadout:` filter now matches partial loadout names -- use `inloadout:"pvp"` for items in loadouts where "pvp" is in the loadout's name.
-* Fashion is now applied to items in a loadout.
+* If your loadout includes ornaments, items are shown as if they had the loadout applied in the loadout page and loadout editor.
 * You can now change the Aeon sect mod through the item popup.
 * You can now edit your equipped Emotes from DIM. You can't add them to loadouts... yet.
 * There's a good chance craftable items will display correctly in DIM. No promises though.